+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 All notable changes to this project will be documented in this file. This project adheres to Semantic Versioning.
 
+### 5.4.0
+* Add functionality to send disabled products to Nosto
+* Add ttl for nosto_product_cache
+* Refactor BuilderTrait into service classes
+
 ### 5.3.2
 * Improve exception during product build by passing previos exception